Rename some more libs
