+++ src/bot/commands/PlayerInfo.ts
@@ -3,7 +3,7 @@ import Command from "../structs/Command";
 import players from "../../database/schemas/players";
 import * as bot from "../bot";
 import { translateToAnsi } from "../../libs/chat-color";
-import { getProfileByUsername } from "../../libs/mojang";
+import { getProfileByUsername } from "../../libs/game-profiles";
 import { capitalCase } from "change-case";
 export const uuidRegex = /[a-f0-9]{8}(?:-[a-f0-9]{4}){4}[a-f0-9]{8}|[a-f0-9]{8}(?:[a-f0-9]{4}){4}[a-f0-9]{8}/;
